fix: resolve 13 type name collisions in Sprint 1311-1320 headers
- Renamed 13 conflicting types across 18 source/header files to prevent C2011
  redefinition errors when Sprint 1311-1320 headers are co-included with prior
  sprint headers in EngineTestsIncludes.h

  Renamed types:
  - AnimatedFormat -> SampledAnimFormat (AnimatedSequenceSampler)
  - ToneMapOperator -> PQToneMapOp (PQToSDRToneMapper)
  - LTSGateStatus -> LTSValidatorStatus (LTSBuildValidator)
  - ScrubState -> HoverScrubState (HoverScrubController)
  - LatencyPercentiles -> HistogramPercentiles (DecodeLatencyHistogram)
  - DecodeRequest -> WorkerDecodeRequest (ZeroTrustDecodeWorker)
  - WorkerState -> ZTWorkerState (ZeroTrustDecodeWorker)
  - AuditEvent -> EnterpriseAuditEvent (EnterpriseAuditExporter)
  - PrefetchRequest -> AsyncPrefetchItem (AsyncPrefetchQueue)
  - BoundingBox3D -> STEPBoundingBox (STEPBoundingBoxExtractor)
  - DICOMWindowPreset -> AdvancedWindowPreset (DICOMAdvancedDecoder)
  - PressureLevel -> ResponderPressureLevel (MemoryPressureResponder)
  - GateResult/GateVerdict -> GPUGateResult/GPUGateVerdict (GPUDecodePerformanceGate)

- Updated 3 test split files to reference renamed types + fixed field names
- Fixed C4189 unused variable warning in EngineTests_Late.cpp (removed 'dxf')
- BUILD_SUCCESS: 0 errors, 0 warnings, 20/20 steps, EngineTests.exe linked
- TEST_SUCCESS: all 4704 tests pass

Docs updated:
- lessons-learned.md: added §11.8 Sprint 1311-1320 Type Collision Crisis
- lessons-learned.md: removed duplicate §11 section (structural bug)
- copilot-instructions.md: added rule #24 pre-sprint collision check (MANDATORY)
- gh-tracking-log.md: added v35.3.0 build-fix session record

+++ Engine/Decoders/FITSZScaleStretch.cpp
@@ -43,7 +43,7 @@ ZScaleLimits FITSZScaleStretch::ComputeLimits(
     const float slope = (n > 1) ? (vmax - vmin) / static_cast<float>(n - 1u) : 1.0f;
 
     const float halfRange = (opts.contrast > 0.0f)
-        ? slope * static_cast<float>(n) * 0.5f / opts.contrast
+        ? slope * static_cast<float>(n) * 0.5f / static_cast<float>(opts.contrast)
         : (vmax - vmin) * 0.5f;
 
     lim.z1    = std::max(static_cast<double>(vmin),
